Fix private route not rendering
Change the behavior of PrivateRoute component to verify if the user is
authenticated

When the component verify if user is signed in, it needs to consider
the case of the jwt expires. The previous solution was to check if the
local storage value of userId key or the user value of the fetched user
are falsy, but such condition suites more cases than the expected,
which results in an inconsistently behavior of the software. To resolve
it, the component will check only the case where there's a value on
userId key on local storage and the id of user context value is falsy.S