Fix filter multiplying letters in Tags
When the filter is applied, it highlights the letters in common with
the search string, by splitting the tag name with the search string as
separator and using a reducer to add the search string on everytime it
was splitted. But the method split of javascript's array creates an
empty string when it finds the separator, then the solution was to add
the search string when the reducer finds an empty string.