Change context providers name
Currently the provider components are just a function with Provider
as name but any error message within these providers can't show
which provider it is because it uses the function name to tell
the component's name. Thus, explicitly showing which provider is it
by its function name can avoid some headaches when debugging.